Improve module structure, general tidy up
